+++ www/go/base/mail/Message.php
@@ -46,10 +46,10 @@ class Message extends \Swift_Message{
 	
 	public function __construct($subject = null, $body = null, $contentType = null, $charset = null) {
 		parent::__construct($subject, $body, $contentType, $charset);
-		
-//		$headers = $this->getHeaders();
 
-//		$headers->addTextHeader("User-Agent", GO::config()->product_name);
+    $headers = $this->getHeaders();
+    $headers->addTextHeader("X-Group-Office-Title", go()->getSettings()->title);
+
 		// See Mailer.php at line 105 for header encoding
 		if(GO::config()->swift_email_body_force_to_base64) {
 			//Override qupted-prinatble encdoding with base64 because it uses much less memory on larger bodies. See also:
@@ -105,7 +105,7 @@ class Message extends \Swift_Message{
   public static function newInstance($subject = null, $body = null,
     $contentType = null, $charset = null)
   {
-    return new self($subject, $body, $contentType, $charset);
+    $msg = new self($subject, $body, $contentType, $charset);
   }
 	
 	/**