+++ www/go/core/mail/Message.php
@@ -23,6 +23,9 @@ class Message extends \Swift_Message {
 		$this->mailer = $mailer;
 		parent::__construct();
 		$this->setFrom(go()->getSettings()->systemEmail,go()->getSettings()->title);
+
+    $headers = $this->getHeaders();
+    $headers->addTextHeader("X-Group-Office-Title", go()->getSettings()->title);
 	}
 
 	/**