+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+- Email: Fixed error in sieve disconnect
+
 06-12-2019 6.4.88
 - Files: New Assistant implementation without webdav requirement. Clients need to be updated!
 - Address book: Added zipcode to text search and as zip: 1234 AB